+++ src/packages/Redact.ts
@@ -1,4 +1,5 @@
 import traverse from 'traverse';
+import * as R from 'ramda';
 import { RedactClass } from '../../types';
 
 const KEYS = [
@@ -46,12 +47,14 @@ class Redact implements RedactClass {
     });
   }
 
-  map(obj: unknown): unknown {
+  map<T>(obj: T): T {
     // eslint-disable-next-line @typescript-eslint/no-this-alias
     const self = this;
     // eslint-disable-next-line array-callback-return
-    return traverse(obj).map(function _nameless(value): void {
-      if (self.key(this.key || '')) {
+    return traverse<T>(obj).map(function _nameless(value: unknown): void {
+      const key: string | null = R.pathOr(null, ['key'], this);
+
+      if (key && self.key(key)) {
         this.update(self.redacted);
         return;
       }